refactor(admin-ui-react): remove dead code
