+++ packages/admin-ui-react/src/index.tsx
@@ -8,8 +8,6 @@ export * from './hooks/useStylesheet'
 export * from './hooks/useMediaQuery'
 export * from './hooks/useResponsiveValue'
 export * from './createSystem'
-export * from './themeMode'
-export * from './hooks/useStoredState'
 export * from './context'
 export { IconContainer } from './createIcons'
 export * from './system'
